==> NSCLC2/01_NSCLC2_clinical_analysis.Rmd <==
---
title: "NSCLC2_clinical_analysis"
author: "Francesco Massaini"
date: "2024-06-11"
output: html_document
---

# Imports
```{r Imports}
suppressMessages(library(dplyr)) 
suppressMessages(library(tibble))
suppressMessages(library(tidyr))
suppressMessages(library(org.Hs.eg.db)) # Homo sapiens db
suppressMessages(library(stringr))
suppressMessages(library(RColorBrewer)) # for plot
suppressMessages(library(pheatmap)) # for heatmaps
suppressMessages(library(ADImpute)) # This package provides functions to compute TPM
suppressMessages(library(ggplot2))
suppressMessages(library(FactoMineR))
suppressMessages(library(factoextra))
suppressMessages(library(survival))
suppressMessages(library(ranger))
suppressMessages(library(ggfortify))
suppressMessages(library(rlang))
suppressMessages(library(purrr))
suppressMessages(library(rio))
suppressMessages(library(uwot))
suppressMessages(library(DESeq2))
```


# Functions
```{r Functions}
source("../Functions.R")
```

# Load clinical
```{r}
clinical = read.csv("../../IMMUCAN_data/NSCLC2/01_Clinical_Data/Daniel/NSCLC2_for_R_update_2024.csv") %>%
  dplyr::select(immucan_id, stage, death, neo_adjuvant_treatment, adjuvant_treatment, simple_histology, long_survivors, TIL_score, side_localization, gender) %>%
  column_to_rownames(var = "immucan_id") 
```

# Fisher/chi tests
```{r}

for (feat1 in colnames(clinical)){
  clinical_features = colnames(clinical)[colnames(clinical) != feat1]
  for (feat2 in clinical_features){
    selected = dplyr::select(clinical, feat1 , feat2)
    contingency_table = table(selected)
    chi_result <- chisq.test(as.matrix(contingency_table))
    if (!is.na(chi_result$p.value)) {
      if (chi_result$p.value < 0.95) {
        print(paste(feat2, chi_result$p.value))
        print(contingency_table)
        mosaicplot(contingency_table + 0.1,
                   main = "Mosaic plot",
                   sub = paste("P-value:", format(chi_result$p.value, digits = 2)),
                   color = TRUE
                  )
        } else{
      cat(feat2, "No Significance\n")
        }
      }
    }
    
}  
```




# Kaplan-Meier
```{r}
clinical_surv = read.csv("../../IMMUCAN_data/NSCLC2/01_Clinical_Data/Daniel/NSCLC2_for_R_update_2024.csv") %>%
  dplyr::select(immucan_id, time_to_event, stage, death, neo_adjuvant_treatment, adjuvant_treatment, simple_histology, long_survivors, TIL_score, side_localization, gender) %>%
  mutate(death = ifelse(death == "Yes", 1, 0)) %>%
  column_to_rownames(var = "immucan_id") 

km_trt_fit <- survfit(Surv(time_to_event, death) ~ 1, data=clinical_surv)
autoplot(km_trt_fit)

km_trt_fit <- survfit(Surv(time_to_event, death) ~ gender, data=clinical_surv)
autoplot(km_trt_fit)

km_trt_fit <- survfit(Surv(time_to_event, death) ~ TIL_score, data=clinical_surv)
autoplot(km_trt_fit)

km_trt_fit <- survfit(Surv(time_to_event, death) ~ adjuvant_treatment, data=clinical_surv)
autoplot(km_trt_fit)

km_trt_fit <- survfit(Surv(time_to_event, death) ~ simple_histology, data=clinical_surv)
autoplot(km_trt_fit)
```
